Calculer l'évolution à partir d'une table agrégée

Cette fonction calcule l'évolution d'une variable quantitative sur la base d'une table agrégée. Elle transforme les données en un format large et calcule le pourcentage d'évolution entre les années.

Usage

calculer_evolution_from_table_agrege(
  table_agrege,
  var_evolution,
  var_croisement,
  var_croisement_relative,
  var_quanti_selec
)

Arguments

table_agrege

Un data.frame contenant les données agrégées.

var_evolution

Un caractère spécifiant la variable d'évolution.

var_croisement

Un vecteur de caractères spécifiant les variables de croisement.

var_croisement_relative

Un vecteur de caractères spécifiant les variables de croisement relatives.

var_quanti

Un caractère spécifiant la variable quantitative pour laquelle l'évolution est calculée.

Value

Un data.frame contenant les colonnes de croisement, les colonnes de croisement relative, et la colonne d'évolution calculée.