Creer un graphique a barres avec ggplot2 et le sauvegarder automatiquement

Cette fonction genere un graphique a barres a partir d'un jeu de donnees en utilisant ggplot2. Le graphique est automatiquement sauvegarde dans un repertoire 'output' avec un nom de fichier genere a partir du titre du graphique.

Usage

creer_graphique_bar(
  data,
  var_x,
  var_y,
  var_fill,
  lab_x,
  lab_y,
  titre,
  labels_fill = NULL,
  soustitre = NULL,
  color_theme = c("#FF4858", "#1B7F79", "#00CCC0", "#72F2EB", "#747F7F"),
  param_position = "stack",
  save = FALSE
)

Arguments

data

data.frame contenant lesgit push -u origin main donnees a representer.

var_x

Chaîne de caracteres specifiant le nom de la variable pour l'axe des abscisses dans data.

var_y

Chaîne de caracteres sgipecifiant le nom de la variable pour l'axe des ordonnees dans data.

var_fill

Chaîne de caracteres specifiant le nom de la variable de remplissage (couleur) dans data.

lab_x

Chaîne de caracteres pour le label de l'axe des abscisses.

lab_y

Chaîne de caracteres pour le label de l'axe des ordonnees.

titre

Chaîne de caracteres pour le titre du graphique.

labels_fill

Vecteur de chaînes de caracteres pour les labels de la legende des couleurs.

soustitre

(Optionnel) Chaîne de caracteres pour le sous-titre du graphique. Par defaut : NULL.

color_theme

Vecteur de codes couleurs hexadecimaux pour le graphique. Par defaut : palette de couleurs predefinie.

param_position

Chaîne de caracteres specifiant la position des barres ("stack", "dodge", etc.). Par defaut : "stack".

save

booleen permettant de choisir si l'on veut sauvegarder ou non.

Value

Un objet ggplot2 representant le graphique a barres.

Details

La fonction cree un graphique a barres en utilisant ggplot2 et le sauvegarde automatiquement dans un repertoire 'output'. Le nom du fichier est genere a partir du titre du graphique en remplaçant les espaces par des underscores.

Les couleurs utilisees pour le remplissage sont determinees par color_theme, et les labels de la legende sont specifies par labels_fill. Si le repertoire 'output' n'existe pas, il sera cre automatiquement.

Examples

if (FALSE) { # \dontrun{
creer_graphique_bar(
  data = taux_global_long,
  var_x = "dep_EP",
  var_y = "percentage",
  var_fill = "region",
  lab_x = "Departement",
  lab_y = "Pourcentage",
  titre = "Structure globale de detention",
  labels_fill = c("filiales 971", "filiales 972")
)
} # }