Gerer une demande d'agregation a partir d'un vecteur de parametres

Cette fonction traite une demande specifique d'agregation en utilisant un vecteur de parametres. Elle recupere les informations necessaires, calcule l'agregat en fonction des parametres specifies et retourne le resultat.

Usage

gerer_une_demande(vecteur_demande, metadata = NULL, for_ollama = FALSE)

Arguments

vecteur_demande

Un vecteur contenant les parametres de la demande. Les elements du vecteur doivent être dans l'ordre suivant :

id_demande : Identifiant de la demande.

table : Nom de la table de donnees a utiliser.

var_croisement : Variables de croisement, separees par des tirets.

var_croisement_relative : Variable de croisement relative (unique).

var_quanti : Variable quantitative sur laquelle effectuer les agregations.

fonctions_agregations : Fonctions d'agregation a appliquer, separees par des tirets.

condition : Condition a appliquer pour filtrer les donnees (expression R sous forme de texte).

type_output : Type de sortie souhaite (par exemple, "table", "graph").

nom_fichier_xls : Nom du fichier Excel de sortie.

nom_onglet : Nom de l'onglet dans le fichier Excel.

titre : Titre pour la sortie (par exemple, titre du graphique).

unite : Unites de mesure, separees par des tirets.

for_ollama

Logical. Si TRUE, retourne uniquement la table agregee sans traitement supplementaire.

Value

Selon les parametres :

Un data.frame contenant le resultat de l'agregation si type_output = "table" ou for_ollama = TRUE

Un objet ggplot si type_output = "graphique"

Details

La fonction effectue les etapes suivantes :

Decompose le vecteur_demande en variables individuelles.

Charge la table de donnees correspondante en utilisant get().

Extrait les variables de croisement et les fonctions d'agregation.

evalue la condition fournie pour filtrer les donnees si necessaire.

Calcule l'agregat en appelant la fonction calculer_agregat_sur_croisement() avec les parametres extraits.

See also

Examples

if (FALSE) { # \dontrun{
# Exemple d'utilisation de gerer_une_demande
# Supposons que vecteur_demande soit un vecteur correctement formate
vecteur_demande <- c(
  id_demande = 1,
  table = "donnees_ventes",
  var_croisement = "region-produit",
  var_croisement_relative = "region",
  var_quanti = "chiffre_affaires",
  fonctions_agregations = "sum-mean",
  condition = "annee == 2023",
  type_output = "table",
  nom_fichier_xls = "rapport_ventes.xlsx",
  nom_onglet = "Synthese",
  titre = "Rapport des ventes 2023",
  unite = "euros"
)
resultat <- gerer_une_demande(vecteur_demande)
} # }